lesson-05-api-integration/demo/requirements.txt
google-adk
google-cloud-secret-manager
google-maps-places
